+++ packages/sdk/test/devnet/dev_sale_fa2.ts
@@ -12,6 +12,8 @@ export async function dev_sale_fa2() {
     edsk: sellerEdsk,
     contract: "KT1NWdwVA8zq5DDJTKcMkRqWYJcEcyTTm5WK",
     amount: 1,
+    metadata: `{"": "ipfs://QmQ4x5BR7ecGVjyhZ7o87m2rPgzp8sBzxFbM4gtHiQQ6ay"}`,
+    royalties: `{"tz1Mxsc66En4HsVHr6rppYZW82ZpLhpupToC": 1000}`,
     is_dev: true
   })
   console.log('mintedItemId', mintedItemId)