+++ packages/sdk/test/testnet/bundle_sale_fa12.ts
@@ -4,14 +4,17 @@ import {awaitItem} from "../common/utils";
 
 export async function bundle_sale_fa12() {
   console.log("--------------------")
-  console.log("Running bundle_sale_fa12 test")
+  console.log("Running dev_bundle_sale_fa12 test")
   console.log("--------------------")
   const sellerEdsk = "edskRqrEPcFetuV7xDMMFXHLMPbsTawXZjH9yrEz4RBqH1D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj"
   const buyerEdsk = "edskS4QxJFDSkHaf6Ax3ByfrZj5cKvLUR813uqwE94baan31c1cPPTMvoAvUKbEv2xM9mvtwoLANNTBSdyZf3CCyN2re7qZyi3"
 
   const mintedItemId_0 = await testScript('mint', {
     edsk: sellerEdsk,
-    contract: "KT1EreNsT2gXRvuTUrpx6Ju4WMug5xcEpr43"
+    contract: "KT1Uke8qc4YTfP41dGuoGC8UsgRyCtyvKPLA",
+    metadata: `{"": "ipfs://QmQ4x5BR7ecGVjyhZ7o87m2rPgzp8sBzxFbM4gtHiQQ6ay"}`,
+    royalties: `{"tz1Mxsc66En4HsVHr6rppYZW82ZpLhpupToC": 1000}`,
+    amount: 100
   })
   console.log('mintedItemId_0', mintedItemId_0)
 
@@ -19,7 +22,10 @@ export async function bundle_sale_fa12() {
 
   const mintedItemId_1 = await testScript('mint', {
     edsk: sellerEdsk,
-    contract: "KT1EreNsT2gXRvuTUrpx6Ju4WMug5xcEpr43"
+    contract: "KT1Uke8qc4YTfP41dGuoGC8UsgRyCtyvKPLA",
+    metadata: `{"": "ipfs://QmQ4x5BR7ecGVjyhZ7o87m2rPgzp8sBzxFbM4gtHiQQ6ay"}`,
+    royalties: `{"tz1Mxsc66En4HsVHr6rppYZW82ZpLhpupToC": 1000}`,
+    amount: 100
   })
   console.log('mintedItemId_1', mintedItemId_1)
 
@@ -30,7 +36,9 @@ export async function bundle_sale_fa12() {
     edsk: sellerEdsk,
     item_id: `${mintedItemId_0},${mintedItemId_1}`,
     sale_type: AssetTypeV2.FA12,
-    ft_contract: "KT1WsXMAzcre2MNUjNkGtVQLpsTnNFhBJhLv"
+    ft_contract: "KT1WsXMAzcre2MNUjNkGtVQLpsTnNFhBJhLv",
+    qty: 1,
+    amount: 0.000002,
   })
   console.log('sellOrder', sellOrder)
 
@@ -40,7 +48,8 @@ export async function bundle_sale_fa12() {
     owner: "tz1Mxsc66En4HsVHr6rppYZW82ZpLhpupToC",
     sale_type: AssetTypeV2.FA12,
     ft_contract: "KT1WsXMAzcre2MNUjNkGtVQLpsTnNFhBJhLv",
-    amount: 0.02
+    qty: 1,
+    amount: 0.000002,
   })
   console.log('buyOrder', buyOrder)
 }