+++ packages/sdk/test/testnet/deploy_nft.ts
@@ -6,11 +6,9 @@ export async function deploy_nft() {
   console.log("Running deploy_nft test")
   console.log("--------------------")
   const sellerEdsk = "edskRqrEPcFetuV7xDMMFXHLMPbsTawXZjH9yrEz4RBqH1D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj"
-  const buyerEdsk = "edskS4QxJFDSkHaf6Ax3ByfrZj5cKvLUR813uqwE94baan31c1cPPTMvoAvUKbEv2xM9mvtwoLANNTBSdyZf3CCyN2re7qZyi3"
 
   const deployNftResult = await testScript('deploy_nft', {
-    edsk: sellerEdsk,
-    is_dev: true
+    edsk: sellerEdsk
   })
   console.log('deployNftResult', deployNftResult)