+++ packages/sdk/test/testnet/get_auction.ts
@@ -7,7 +7,7 @@ export async function get_auction() {
   const sellerEdsk = "edskRqrEPcFetuV7xDMMFXHLMPbsTawXZjH9yrEz4RBqH1D6H8CeZTTtjGA3ynjTqD8Sgmksi7p5g3u5KUEVqX2EWrRnq5Bymj"
   const buyerEdsk = "edskS4QxJFDSkHaf6Ax3ByfrZj5cKvLUR813uqwE94baan31c1cPPTMvoAvUKbEv2xM9mvtwoLANNTBSdyZf3CCyN2re7qZyi3"
 
-  const mintedItemId = "KT1EreNsT2gXRvuTUrpx6Ju4WMug5xcEpr43:247"
+  const mintedItemId = "KT1Uke8qc4YTfP41dGuoGC8UsgRyCtyvKPLA:247"
   const auction = await testScript('get_auction', {
     edsk: sellerEdsk,
     item_id: mintedItemId